+++ src/main/java/org/realityforge/replicant/server/ee/rest/ReplicantPollResource.java
@@ -1,5 +1,6 @@
 package org.realityforge.replicant.server.ee.rest;
 
+import java.util.Date;
 import java.util.Iterator;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
@@ -24,6 +25,7 @@ import javax.ws.rs.container.AsyncResponse;
 import javax.ws.rs.container.ConnectionCallback;
 import javax.ws.rs.container.Suspended;
 import javax.ws.rs.container.TimeoutHandler;
+import javax.ws.rs.core.Response;
 import org.realityforge.replicant.shared.transport.ReplicantContext;
 
 /**
@@ -90,7 +92,7 @@ public class ReplicantPollResource
       final String data = _source.poll( sessionID, rxSequence );
       if ( null != data )
       {
-        response.resume( data );
+        resume( response, data );
       }
       else
       {
@@ -99,7 +101,7 @@ public class ReplicantPollResource
     }
     catch ( final Exception e )
     {
-      response.resume( e );
+      resume( response, e );
     }
   }
 
@@ -125,7 +127,32 @@ public class ReplicantPollResource
   private void doTimeout( final AsyncResponse response )
   {
     doDisconnect( response );
-    response.resume( "" );
+    resume( response, "" );
+  }
+
+  private static void resume( final AsyncResponse asyncResponse, final Object message )
+  {
+    asyncResponse.resume( toResponse( message ) );
+  }
+
+  private static Response toResponse( final Object message )
+  {
+    final Date now = new Date();
+    return Response.ok().
+      // set create date to current timestamp
+        header( "Date", now.getTime() ).
+      // set modify date to current timestamp
+        header( "Last-Modified", now.getTime() ).
+      // set expiry to back in the past (makes us a bad candidate for caching)
+        header( "Expires", 0 ).
+      // HTTP 1.0 (disable caching)
+        header( "Pragma", "no-cache" ).
+      // HTTP 1.1 (disable caching of any kind)
+        // HTTP 1.1 'pre-check=0, post-check=0' => (Internet Explorer should always check)
+        //Note: no-store is not included here as it will disable offline application storage on Firefox
+        header( "Cache-control", "no-cache, must-revalidate, pre-check=0, post-check=0" ).
+      entity( message ).
+      build();
   }
 
   /**
@@ -202,13 +229,13 @@ public class ReplicantPollResource
             final String data = _source.poll( request.getSessionID(), request.getRxSequence() );
             if ( null != data )
             {
-              request.getResponse().resume( data );
+              resume( request.getResponse(), data );
               iterator.remove();
             }
           }
           catch ( final Exception e )
           {
-            request.getResponse().resume( e );
+            resume( request.getResponse(), e );
             iterator.remove();
           }
         }